requirements.txt
# Exo Windows Porting - Python 依赖清单
# ======================================
# 核心依赖

pydantic>=2.0.0
rich>=13.0.0
typer>=0.9.0

# P2P 网络层
pyzmq>=25.0.0
libp2p>=0.1.0
zeroconf>=0.96.0

# LLM Backend (llama.cpp — single-node path)
llama-cpp-python>=0.2.87
numpy>=1.24.0
diskcache>=5.6.1

# Distributed inference (pipeline parallelism across multiple nodes)
# Install the correct torch build for your hardware:
#   CUDA:  pip install torch --index-url https://download.pytorch.org/whl/cu121
#   ROCm:  pip install torch --index-url https://download.pytorch.org/whl/rocm6.1
#   CPU:   pip install torch --index-url https://download.pytorch.org/whl/cpu
torch>=2.1.0
transformers>=4.40.0
accelerate>=0.29.0
safetensors>=0.4.0

# Web Dashboard & API
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
websockets>=12.0
jinja2>=3.1.0

# 测试
pytest>=7.4.0
pytest-asyncio>=0.23.0
httpx>=0.26.0
coverage>=7.0.0

# 开发工具
black>=24.1.0
ruff>=0.3.0
mypy>=1.8.0
pre-commit>=3.6.0
pylint>=3.0.0

# 可选依赖 (ROCm GPU)
# llama-cpp-python[rocm]>=0.2.87  # AMD ROCm 支持

# 可选依赖 (CUDA GPU)  
# llama-cpp-python[cuda]>=0.2.87  # NVIDIA CUDA 支持
